+++ grade.h
@@ -1,0 +1,5 @@
+#include <vector>
+
+double grade (double midterm, double final, double homework);
+double grade (double midterm, double final, const std::vector<double>& hw);
+double grade (const Student_info& s);
