+++ median.h
@@ -1,0 +1,3 @@
+#include <vector>
+
+double median (std::vector<double> vec);
