+++ mod.cpp
@@ -1,3 +1,7 @@
+#include "median.h"
+#include "student_info.h"
+#include "grade.h"
+
 #include <algorithm>
 #include <iomanip>
 #include <iostream>
@@ -21,72 +25,8 @@ using std::string;
 using std::vector;
 
 
-struct Student_info {
-	string name;
-	double midterm;
-	double final;
-	vector<double> homework;
-};
 
 
-//==============================================================================
-double median (vector<double> vec) {
-	size_t size = vec.size();
-	if (size == 0) throw domain_error("Вектор пуст! Вычисление медианы невозможно.");
-
-	sort(vec.begin(), vec.end());
-	size_t mid = size / 2;
-
-	return size % 2 == 0 ? (vec[mid] + vec[mid-1]) / 2 : vec[mid];
-}
-
-
-//==============================================================================
-double grade (double midterm, double final, double homework) {
-	return (0.2 * midterm + 0.4 * final + 0.4 * homework);
-}
-
-
-//==============================================================================
-double grade (double midterm, double final, const vector<double>& hw) {
-	if (hw.size() == 0) throw domain_error("Отсутствуют оценки за самостоятельные работы!");
-
-	return grade(midterm, final, median(hw));
-}
-
-
-//==============================================================================
-double grade (const Student_info& s) {
-	return grade(s.midterm, s.final, s.homework);
-}
-
-
-//==============================================================================
-istream& read_hw (istream& in, vector<double>& hw) {
-	if (in) {
-		hw.clear();
-		
-		for (double x; in >> x;) hw.push_back(x);
-
-		in.clear();
-	}
-	return in;
-}
-
-
-//==============================================================================
-istream& read (istream& is, Student_info& s) {
-	is >> s.name >> s.midterm >> s.final;
-	read_hw(is, s.homework);
-
-	return is;
-}
-
-
-//==============================================================================
-bool compare (const Student_info& x, const Student_info& y) {
-	return x.name < y.name;
-}
 
 
 //==============================================================================
