+++ tests/studs.txt
@@ -7,7 +7,7 @@
 Русаков 7.2 8.7 8.8 5.4 5.5 8.2 6.9 8.7
 Фомина 9.0 9.6 9.9 9.9 10.0 8.1 9.7 9.7
 Сапожников 8.1 9.7 9.9 6.7 4.0 9.0 7.0 9.6
-Петренко 4.3 9.8 9.6 7.9 10.0 8.2 9.7 9.6
+Петренко 1 2 3 4 5 6 7 8 9 10
 
 
 Кузнецов 6.7 7.2 7.3 4.0 0 7.8 5.5 7.0
